backend/requirements.txt
# Django Core
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-otp==1.3.0
qrcode==7.4.2
cryptography==41.0.7
PyJWT==2.8.0

# Celery & Redis
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# API Documentation
drf-spectacular==0.27.0

# Environment Variables
python-decouple==3.8

# WSGI Server
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==23.12.1
flake8==7.0.0
isort==5.13.2

# Monitoring
sentry-sdk==1.39.1

# File Handling
openpyxl==3.1.2

# SMS (example, replace with actual provider)
twilio==8.11.1
